Adding experimental support for not sampling into suspended threads. Cleaned up a bunch of warnings and misc build issues.

+++ SlimTuneCLR/PerfCounter.cpp
@@ -62,7 +62,7 @@ unsigned int PerfCounter::AddCounterRaw(const std::wstring& counterPath)
 	PDH_HCOUNTER counter;
 	PdhAddCounter(m_query, counterPath.c_str(), NULL, &counter);
 	m_counters.push_back(counter);
-	return m_counters.size();
+	return static_cast<unsigned int>(m_counters.size());
 }
 
 unsigned int PerfCounter::AddInstanceCounter(const std::wstring& objectName, const std::wstring& counterName)
@@ -71,7 +71,7 @@ unsigned int PerfCounter::AddInstanceCounter(const std::wstring& objectName, con
 	std::wstring fullCounterName = (boost::wformat(L"\\%1%(%2%)\\%3%") % objectName % m_instName % counterName).str();
 	PdhAddCounter(m_query, fullCounterName.c_str(), NULL, &counter);
 	m_counters.push_back(counter);
-	return m_counters.size();
+	return static_cast<unsigned int>(m_counters.size());
 }
 
 void PerfCounter::Update()
